Hans Rosling - data visualisation project
The visualization has three elements in it as follows,
1. The bubble plot representation.
2. The checkbox list.
3. Bar graph of 12 pillars for the selected two countries.

In the first part, we have created a bubble plot wherein the x-axis represents the GDP, the Y-axis represents the Global Competitive Index and the radius of the circle represents the total population of the country.
The GDP has been scaled with log values whereas the Global Competitive index is scaled linearly.
I have used well-defined ticks which are labeled properly. As log scale was used in x-axis the ticks were not equal in distance hence, with the help of array and tick format; that is .1s I made sure they are well presented to the user.
‘Mouse hover’ functionality is built wherein, by hovering a particular circle we can see the name of the country and two lines are mapped on x and y-axis to show the values with respect to the hovered circle.
The play button and slider is placed just below the graph that can be used to see the transition. Also, the slider is labeled with years allows the user to click on the particular year and bubbles are transitioned accordingly. The year for which the data is represented is displayed at the back of the visualization.
Colour for the bubbles are selected on the basis of the Forum classification field.
The legend is created towards the left side of the graph, which represents the color used for specific regions.
In the second part, we have created trace functionality.
when the visualization is in the pause mode we can click on the particular circle (country) to find out the trace of that circle (country) for all the years.
A checkbox is provided to the user to select the single or two countries for comparisons in the multivariate graph. When the two countries are selected the bar graph for the two countries are created below the bubble chart all the 12 pillars as mentioned in the database. This bar graph varies with respect to the year as the play button is clicked along with the bubble plot. The text box is provided in the bar graph which represents the name of the country whose data is being displayed.
Also, the color of the bar chart is changed with respect to its region which is represented in the bubble graph.